moves Interval Form into the App to clear up distinction between "o- objects" and "c- components" in the BEMIT sense
as per the terrific

https://csswizardry.com/2015/03/more-transparent-ui-code-with-namespaces
/

which really helps think through your UI architecture in terms of scope
of impact that a piece of UI code has.

Objects (o-):
Global base UI. This is atoms and molecules of your app: the basic
building blocks that everything else is based on — changing these will
have significant impact on the rest of your code.

Components (c-):
Local specific UI: the implementation UI for a particular feature in
the app. Likely a component will be used around once (maybe 2 or 3
times) in your app. It is relatively safe to change these and only have
localized impact on your app.